xen: Don't perform SMP setup.
This is not needed and in case one want to start a guest with fewer vCPU
than the maximum, the function will fail (infinite loop).

It's a "backport" of 5dbf1732940f94771e0b3c45c0960940276bc263. The original
commit is rather involved and relies on other infrastructure changes. This
replicates the functional affect.

Signed-off-by: Anthony PERARD <anthony.perard@citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>
[ ijc -- expanded "backport" comment ]